Split the remaining oversized mobile and pane screens
Same treatment as the desktop screens: files cut along the seams they already
had, no change to what any composable does.

  MobileTerminalView    988 -> 454 + chrome 296 + ai bar 224 + password sheet 109
  MobileFilesView       828 -> 267 + pane 291 + chrome 265 + mock 103
  MobileDesignApp       817 -> 326 + chrome 377 + session actions 140
  TerminalView          752 -> 367 + panes 426

Two more already-dead declarations go rather than keep their suppressions:
MobileRoutePlaceholder, and NoSessionMobileFilesView's unused `mono` parameter.
